Re-work ProtocolsModelComponent constructor
Update NewProtocolsModelComponent() to allocate the Component and
Model objects up front, before Run() gets called. To do this we need
the component name to be provided up front, so this is moved from
Run().

This allows us to easily support multiple RPCs by setting each
directly in the Model object, rather than maintaining the previous
indirection in ProtocolsModelComponent.

GetRPCName() is removed as it no longer makes sense.

This commit is therefore an API breaking change.

VRVDR-50933