Correctly escape hyphen character in community-list string patterns
The hyphen ('-') character in the string patterns of the community-list*
and extcommunity-list* typedefs are treated as if they define a range of
characters to include in the class, not as if they represent a literal
hyphen character which I believe in the intent.

This leads to an interesting issue where seemingly the numeral range
defined earlier in the character class ('0-9') is overridden, or is no
longer valid.

E.g. Attempting to validate the values such as 'a1', '1a', '11a', 'a11'
and similar against node which use those typedefs fails.

This commit escapes the affected hyphen characters correctly which
results in the above examples being accepted.

Resolves: VRVDR-51993