yang: allow MPLS on bonding interfaces
In general, anything that is allowed to be configured on a dataplane
interface should be allowed on a bonding interface since they are
handled in the same way and users have an expectation that they can be
used interchangeably.

So allow MPLS (particularly LDP) to be configured on bonding and
bonding vif interfaces.

Fixes: VRVDR-53431